Fix dtype mangling
The combination of torch creating tensors with a default type (usually of a float variety) and type promotion means that tensors containing integers become converted to floats.

Here we are more explicity about dtypes to avoid the create of arrays with the torch default dtype, but rather base the type on the relevant tensors.